requirements.txt
python-dotenv
pyttsx3
pyaudio
openai-whisper
google-generativeai
keyboard

protobuf